requirements.txt
numpy
folium
geopy
flask-cors
eli5
-e .